Change logic for deciding when to publish
Previously, I was only publishing on a timer. If the timer when off and
there was at least one measurement in the queue, then I tried publishing
it. To change the batch size, I would change the timer. Now, I leave the
timer the same (1 second) and change a new configuration option,
UPLOAD_BATCH_SIZE. This means a publish will only happen if the amount
of measurements is greater than the batch. Much better.